requirements.txt
streamlit==1.3.0
numpy==1.20.3
onnxruntime==1.10.0
pillow==8.2.0
